+++ R/table-notes.R
@@ -10,11 +10,12 @@
 #' @param output_file the name of the output file where the table text will be
 #' saved; the file name will be included in the notes in the table footer; see
 #' also [st_files()]
+#' @param ... not used
 #'
 #'@export
 tab_notes <- function(notes, escape_fun = tab_escape,
                       note_config = noteconf(type = "tpt"),
-                      r_file = NULL, output_file = NULL) {
+                      r_file = NULL, output_file = NULL, ...) {
 
   assert_that(is.noteconfig(note_config))